# backend/openapi_static_backup.json
{"openapi":"3.1.0","info":{"title":"SAMBA Simulator API","version":"1.0.0"},"paths":{"/healthz":{"get":{"tags":["health"],"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"Login via form-data (username -> email, password).\nRetorna:\n- access_token (JWT) no body\n- refresh_token no body (para testes) e via cookie HttpOnly","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh":{"post":{"tags":["auth"],"summary":"Refresh Token","description":"Troca um refresh vÃ¡lido por NOVOS tokens (access + refresh).\n- Aceita refresh no cookie OU no body (Swagger-friendly).\n- Aplica rotaÃ§Ã£o: o refresh antigo Ã© revogado.","operationId":"refresh_token_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/RefreshIn"},{"type":"null"}],"title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","description":"Logout:\n- logout_all=True: revoga TODOS os refresh do usuÃ¡rio (logout global)\n- logout_all=False: revoga apenas o refresh atual (cookie/body/header)","operationId":"logout_auth_logout_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"HTTPBearer":[]}]}},"/auth/users":{"post":{"tags":["auth"],"summary":"Create User","operationId":"create_user_auth_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/disciplines/":{"get":{"tags":["disciplines"],"summary":"List Disciplines","operationId":"list_disciplines_disciplines__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DisciplineOut"},"type":"array","title":"Response List Disciplines Disciplines  Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["disciplines"],"summary":"Create Discipline","operationId":"create_discipline_disciplines__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DisciplineCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DisciplineOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/pdf/exams/{exam_id}/pdf/generate":{"post":{"tags":["pdf"],"summary":"Generate Pdfs","description":"Gera PDFs:\n- Por turma: informar class_id\n- Por aluno: informar student_id\n- logo_path: opcional (caminho local no servidor)","operationId":"generate_pdfs_pdf_exams__exam_id__pdf_generate_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"exam_id","in":"path","required":true,"schema":{"type":"integer","title":"Exam Id"}},{"name":"class_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Class Id"}},{"name":"student_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Student Id"}},{"name":"logo_path","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Logo Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pdf/exams/{exam_id}/pdf/download":{"get":{"tags":["pdf"],"summary":"Download Pdf","description":"Download:\n- Por aluno: informar student_id e type = booklet|answer_sheet\n- Por turma: (modo simples) retorna um ZIP no futuro; por enquanto,\n  gere e baixe individualmente (ou use os paths retornados).","operationId":"download_pdf_pdf_exams__exam_id__pdf_download_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"exam_id","in":"path","required":true,"schema":{"type":"integer","title":"Exam Id"}},{"name":"type","in":"query","required":false,"schema":{"enum":["booklet","answer_sheet"],"type":"string","default":"booklet","title":"Type"}},{"name":"class_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Class Id"}},{"name":"student_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"DisciplineCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"DisciplineCreate"},"DisciplineOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"DisciplineOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LogoutIn":{"properties":{"logout_all":{"type":"boolean","title":"Logout All","default":false}},"type":"object","title":"LogoutIn","description":"logout_all=True => revoga todos os refresh do usuÃ¡rio (logout global). "},"RefreshIn":{"properties":{"refresh_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"}},"type":"object","title":"RefreshIn","description":"Para o Swagger: permite enviar refresh no body.\nSe ausente, a rota lÃª do cookie HttpOnly automaticamente."},"TokenOut":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"refresh_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"}},"type":"object","required":["access_token"],"title":"TokenOut","description":"Retorno padrÃ£o de tokens:\n- access_token: JWT curto\n- refresh_token: incluÃ­do para facilitar testes no Swagger\n  (em produÃ§Ã£o, basta ler via cookie HttpOnly)"},"UserCreate":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role"}},"type":"object","required":["name","email","password","role"],"title":"UserCreate"},"UserOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"roles":{"items":{"type":"string"},"type":"array","title":"Roles"}},"type":"object","required":["id","name","email","roles"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}